requirements.txt
langchain>=0.3.25
langchain-gigachat>=0.3.10
langgraph>=0.4.7
python-dotenv>=1.1.0
docx-mailmerge==0.5.0
img2pdf==0.6.1
pytest==8.3.5